fix tmux symlinking

+++ install-scripts/tmux.sh
@@ -1,6 +1,9 @@
 #!/bin/sh
 
 # symlink entire tmux folder into config
+if [[ -e "~/.config/tmux" && ! -L "~/.config/tmux" ]] {
+    exit 0
+}
 ln -sr tmux ~/.config/tmux
 
 git submodule add https://github.com/tmux-plugins/tpm tmux/plugins/tpm
